test_lintel_dbi.pl: Test code that verifies that the Lintel::DBI module is loadable, and exercises the lexical analyzer that extracts SQL statements from documents.  See: nextStatement().  The lexical analyzer knows about standard SQL comments and strings, but does not parse the SQL statement, thus syntax errors will be sent on to the database before they are caught.
Lintel::DBI.pm: Added a TODO noting possible alternatives to
nextStatement() for future investigation.  Commented the code inside
nextStatement(), and repaired a bug in the state transition for
string delimiters.  Simplified the logic so that it is obvious that
nextStatement() is extracting a SQL statement as a token.  Removed the
use of '$&'.

CMakeLists.txt: Removed the old lintel_dbi test in favor of a test modeled
on the perl/test.pl code.