deptool.in: Deptool used to keep track of whether or not it had done an install of a dependency within the same command, but that fails when the dependency is built in a separate command.  This commit changes that to get the time of the last install from its time stamp, and compare that to a project build time when deciding whether to install.
It is still possible that a project will be rebuilt although it didn't 
need to be by being listed after another project which it is not dependent
on but which happens to be newer.  The eventual solution to this is to
build a full graph of the dependencies and their install times, but this
is a quick fix to the missing build problem.