deptool.in: Changed @vcdeps to a reference to allow more parameters to be added to the end of the function call.
The $options{"lastbuild <project>"} holds the last install time of any of
the dependent projects subbuilds.  I don't need to add the subbuild type.
If it is in the dependency list it will be checked by the installation
needed check, which is where that key is filled in, so it is already being
done.

The way that deptool works it can't distinguish between two subtypes
having different file dependencies within its own directory and will
reinstall both cases anyway.  There simply isn't any way that this can
come out wrong until such time as deptool can say that one set of files
is only used for one subbuild type, and another set of files is only
used for the other.