Added deptool-test subdirectory containing the initial work on a deptool test suite.  The current version implements tests 1, 2, 3, 9, 10, 16, and 17 of the test plan.  Embedded within the test is a monotone database with three projects. To modify the source code in that database I have provided a directory 'src' which is a checked out copy of the database branch containing the test projects.
Currently there are two state commented out in test2 related to installing 'c',
and 'a.h'.  The cmake installer fails to copy these files into their respective
install directories if they are already present for unknown reasons.

The tests can be run using the shell scripts 'test[1-30]' located in the deptool-test
subdirectory.  Currently you must cd into the directory to run the test suite.

A minor change was made to deptool.in to allow it to support one character project
names.