* Almost ready for a release, added rules to do the local and remote   host tests.  Need to update to have the code to copy to tesla and   update the NEWS.
checkout.sh: Script to checkout the source and generate the release
  and changelog files.

remote-host-list: list of hosts we test building on along with a comment
  about how they are configured.

test-local.sh: script to build and test on localhost

test-remote.sh: script to build and test on teh various remote hosts.

rules: add a separator line for the new section.  Then lots of different
  targets to do the various phases of the local and remote buidls.

make-dist.sh: Remove the pieces that we have covered in the new dist/rules

deptool.in: New openSUSE has lsb-release, but doesn't have the suitable
  information about the OS in it.  If we don't recognize the OS from 
  lsb-release, fall back to the other recognition methods.