Fix for glitch found on openbsd-4.9-amd64
On that variant, either i386, or 5.0 works, the program options code
was generating a bus error in DataSeries, and it appeared to be because
the program option initialization ran before the boost initialization to
support conversion from doubles to strings.

TODO: Add general TODO to fix this problem overall

ProgramOptions.?pp: Move to having functions to add program option init fns and run them.  The
  adding of them uses the static variable inside a function trick.  Running them happens
  automatically when we parse the command line or a file.

run-check-program-options.sh: add -x to make it easier to see what is failing.