Fixes to get builds working on another machine.
check-changelogs.sh: Script to verify consistency of the changelogs and the version stamp.

make-tar.sh: Bring back the code to make the tar file -- it was removed but turned out to be
  necessary.

rules: fixup $ENV{'HOME'} bits in PB_ROOT.  Get the version from the generated file.  Use the
  new script for checking the changelogs.  make-tar moved to dist.  Split out the rules for
  deb and rpm packages.  Add back in .SECONDARY to preserve the intermediates, e.g. .../setup-vm
  Hide some of the cleanup commands to make the output less messy.

try-fixup-download-errors.sh: For unknown reasons we occasionally get glitches in the package files.
  This script tries to identify those and force the cache to re-fetch the data.